Update wegner.py
Cleaned up the expected arguments of check_all_neighborhood_structures_for_FAS.  (It originally was the code to take a configuration string and check the stem identifications for the standard facial adjacency structure.  Now it starts with a plane graph and lists of roots on outer regions, which generalizes the original routine.)  Small number of changes.